+++ cognit/device_runtime.py
@@ -7,12 +7,11 @@ from cognit.modules._call_queue import CallQueue
 from cognit.modules._logger import CognitLogger
 from threading import Thread
 from typing import Callable
+import signal
+import sys
 
 DEFAULT_CONFIG_PATH = "cognit/config/cognit_v2.yml"
 
-"""
-Class to manage the Device Runtime. It is responsible for offloading functions to the Cognit Frontend
-"""
 class DeviceRuntime:
     
     def __init__(self, config_path=DEFAULT_CONFIG_PATH) -> None:
@@ -23,6 +22,7 @@ class DeviceRuntime:
             config_path (str): Path of the configuration to be applied to access
             the Cognit Frontend
         """
+        
         self.cognit_config = CognitConfig(config_path)
         self.sync_result_queue = SyncResultQueue()
         self.cognit_logger = CognitLogger()
@@ -37,15 +37,28 @@ class DeviceRuntime:
 
         Args:
             init_reqs (dict): requirements to be considered when offloading functions
+
+        Returns:
+            bool: True if the SM thread was launched successfully, False otherwise
         """
+
+        def signal_handler(sig, frame):
+
+            self.stop()
+            sys.exit(0)
+
+        # Handle unexpected app shutdowns
+        signal.signal(signal.SIGINT, signal_handler)
 
         # Check if sm is already running
         if self.sm_thread != None:
+
             self.cognit_logger.error("DeviceRuntime is already running")
             return False
         
         # Check if init_reqs were provided
         if init_reqs == None:
+            
             self.cognit_logger.error("init_reqs not provided")
             return False
         
@@ -54,13 +67,18 @@ class DeviceRuntime:
 
         # State machine initialization
         if self.sm_handler == None:
+
             self.sm_handler = StateMachineHandler(self.cognit_config, self.current_reqs, self.call_queue, self.sync_result_queue)
 
         # Launch SM thread
         try:
+
             self.sm_thread = Thread(target=self.sm_handler.run)
+            self.sm_thread.daemon = True
             self.sm_thread.start()
+
         except Exception as e:
+
             raise Exception(f"DeviceRuntime could not be initialized: {e}")
         
         self.cognit_logger.info("DeviceRuntime initialized")
@@ -75,6 +93,7 @@ class DeviceRuntime:
         """
 
         if self.sm_thread == None:
+
             self.cognit_logger.error("DeviceRuntime is not running")
             return False
 
@@ -92,14 +111,19 @@ class DeviceRuntime:
 
         Args:
             new_reqs (dict): new requirements to be considered when offloading functions
+
+        Returns:
+            bool: True if the requirements were updated successfully, False otherwise
         """
 
         # Check if new_reqs were provided
         if new_reqs == None:
+
             self.cognit_logger.error("new_reqs not provided")
             return False
         
         if self.sm_thread == None:
+
             self.cognit_logger.error("DeviceRuntime is not running")
             return False
 
@@ -108,14 +132,20 @@ class DeviceRuntime:
 
         # Check new_reqs are different from the current ones
         if self.current_reqs == new_reqs:
+
             self.cognit_logger.error("New requirements are the same as the current ones")
             return False
         
+        self.cognit_logger.info(f"Updating requirements from {self.current_reqs} to {new_reqs}")
+        
         # Update the current requirements
         are_updated = self.sm_handler.change_requirements(new_reqs)
+
         if not are_updated:
+
             self.cognit_logger.error("Requirements could not be updated")
             return False
+        
         self.current_reqs = new_reqs
         return True
 
@@ -127,35 +157,48 @@ class DeviceRuntime:
             function (Callable): The target funtion to be offloaded
             callback (Callable): The callback function to be executed after the offloaded function finishes
             params (List[Any]): Arguments needed to call the function
+
+        Returns:
+            bool: True if the function was added to the queue successfully, False otherwise
         """
 
         # Create a Call object
-        call = Call(function=function, fc_lang=FunctionLanguage.PY, mode=ExecutionMode.ASYNC, callback=callback, params=params)
+        call = Call(function=function, fc_lang=FunctionLanguage.PY, mode=ExecutionMode.ASYNC, callback=callback, params=params, timeout=None)
 
         # Add the call to the queue
         if self.call_queue.add_call(call):
+
             self.cognit_logger.debug("Function added to the queue")
             return True
+        
         else:
+
             self.cognit_logger.error("Function could not be added to the queue")
             return False
         
-    def call(self, function: Callable, *params: tuple) -> ExecResponse:
+    def call(self, function: Callable, *params: tuple, timeout: int = None) -> ExecResponse:
         """
         Offloads a function synchronously
 
         Args:
             function (Callable): The target funtion to be offloaded
             params (List[Any]): Arguments needed to call the function
+            timeout (int, optional): Maximum time to wait for the result. Defaults to None.
+
+        Returns:
+            ExecResponse: The response of the offloaded function
         """
 
         # Create a Call object
-        call = Call(function=function, fc_lang=FunctionLanguage.PY, mode=ExecutionMode.SYNC, callback=None, params=params)
+        call = Call(function=function, fc_lang=FunctionLanguage.PY, mode=ExecutionMode.SYNC, callback=None, params=params, timeout=timeout)
 
         # Add the call to the queue
         if self.call_queue.add_call(call):
+
             self.cognit_logger.debug("Function added to the queue")
+
         else:
+
             self.cognit_logger.error("Function could not be added to the queue")
             return None