+++ cognit/modules/_sm_handler.py
@@ -25,6 +25,9 @@ class StateMachineHandler():
 
         Args:
             new_requirements (Scheduling): New requirements to be considered when offloading functions
+
+        Returns:
+            bool: True if the requirements were updated successfully, False otherwise
         """
 
         return self.sm.change_requirements(new_requirements)
@@ -33,6 +36,7 @@ class StateMachineHandler():
         """
         Stop the Device Runtime
         """
+
         self.running = False
     
     def run(self, interval=0.05):
@@ -67,11 +71,13 @@ class StateMachineHandler():
         elif not self.sm.is_ecf_connected():
             self.sm.token_not_valid_ready_2()
         else:
-            if not self.sm.have_requirements_changed():
+            if not self.sm.have_requirements_changed() and not self.sm.is_new_ecf_address_set():
                 self.sm.result_given()
+            elif self.sm.is_new_ecf_address_set() and not self.sm.have_requirements_changed():
+                self.sm.ready_update_ecf_address()
             else:
                 self.sm.ready_update_requirements()
-
+ 
     def handle_init_state(self):
         """
         Handle the init state