+++ cognit/modules/_sync_result_queue.py
@@ -2,15 +2,7 @@ from cognit.models._device_runtime import ExecResponse
 from cognit.modules._logger import CognitLogger
 from threading import Lock, Condition
 
-class Singleton(type):
-    _instances = {}
-
-    def __call__(cls, *args, **kwargs):
-        if cls not in cls._instances:
-            cls._instances[cls] = super(Singleton, cls).__call__(*args, **kwargs)
-        return cls._instances[cls]
-
-class SyncResultQueue(metaclass=Singleton):
+class SyncResultQueue:
 
     def __init__(self):
         self.result = None
@@ -24,13 +16,20 @@ class SyncResultQueue(metaclass=Singleton):
 
         Args:
             result (ExecResponse): Result to be added to the queue
+
+        Returns:
+            bool: True if the result was added successfully, False otherwise
         """
+
         with self.condition:  # Acquire lock for thread safety
+
             if self.result is not None:
                 self.cognit_logger.error("SyncResultsQueue is full. Result will be discarded")
                 return False
+            
             self.result = result
             self.condition.notify()  # Notify any waiting thread that the result is available
+
         return True
 
     def get_sync_result(self) -> ExecResponse:
@@ -40,10 +39,11 @@ class SyncResultQueue(metaclass=Singleton):
         Returns:
             ExecResponse: Result from the queue
         """
+
         with self.condition:  # Acquire lock for thread safety
             while self.result is None:
                 # Wait until there is a result available
-                self.condition.wait()  # Efficiently wait for a result
+                self.condition.wait()
             result = self.result
             self.result = None  # Reset the result
         return result