+++ cognit/test/integration/test_integration_cognit_frontend.py
@@ -14,13 +14,19 @@ COGNIT_CONFIG_PATH = "cognit/test/config/cognit_v2.yml"
 BAD_COGNIT_CONFIG_PATH = "cognit/test/config/cognit_v2_wrong_user.yml"
 
 TEST_REQS_INIT = {
-      "FLAVOUR": "EnergyV2",
-      "GEOLOCATION": "IKERLAN ARRASATE/MONDRAGON 20500"
+    "FLAVOUR": "EnergyTorch",
+    "GEOLOCATION": {
+        "latitude": 43.05,
+        "longitude": -2.53
+    }
 }
 
 REQS_NEW = {
-    "FLAVOUR": "EnergyV2",
-    "GEOLOCATION": "IKERLAN BILBAO 48007"
+    "FLAVOUR": "EnergyTorch",
+    "GEOLOCATION": {
+        "latitude": 43.05,
+        "longitude": -2.53
+    }
 }
 
 # Wrong because "GEOLOCATION" is not defined when "MAX_LATENCY" is defined